Change project structure and add web server
Moved the frontend application into the webapp directory and added
a simple Java web server to the /webserver directory. Added pom files
to build project with maven. Building the project will create
a zip file in the target directory with jar of webserver and build
files of the frontend application.