Add car file to zip and add run script
The car file built from the mi project is added to the zip. Also add
run script to run the application.It runs the webserver to serve
the frontend application and also runs an mi instance with the
built car app. The user must provide the path to the MI_HOME dir
but a seperate copy of it is temporarily created and used to run
the car app. This ensures any changes are isolated from the user's
runtime. The temporary copy is deleted after execution.