Add mi project as a module to root
Previously mvn clean install command should be run in the mi project
directory before running mvn clean package on the root. Now added the
mi project as a module to the root and the required command is executed
in the mi project directory when executing mvn clean package
command on root folder.