Change API calls due to change in API spec
Previously all the details were sent in one api call. Now the details
are sent in the first api call and the file is sent to a different
endpoint seperately. This is to allow for multiple file upload in
the future which can loop through the selected files and upload
each of them one by one to the /upload endpoint of the API.